requirements.txt
langchain
langchain_core
langchain_community
langchain_huggingface
faiss-cpu
pypdf
huggingface_hub
flask
python-dotenv
setuptools
sentence-transformers
langchain_groq